pyproject.toml
[project]
name = "advanced-dl-cv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "pyyaml>=6.0.3",
    "tensorboard>=2.20.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
]
